[TASK] added version 5.2.9

+++ engine/Shopware/Core/sArticles.php
@@ -27,7 +27,6 @@ use Shopware\Bundle\SearchBundle\Sorting\PopularitySorting;
 use Shopware\Bundle\SearchBundle\Sorting\ReleaseDateSorting;
 use Shopware\Bundle\SearchBundle\SortingInterface;
 use Shopware\Bundle\StoreFrontBundle;
-use Shopware\Bundle\StoreFrontBundle\Service\Core\ConfiguratorService;
 use Shopware\Bundle\StoreFrontBundle\Struct\BaseProduct;
 use Shopware\Bundle\StoreFrontBundle\Struct\Product;
 use Shopware\Components\QueryAliasMapper;
@@ -361,6 +360,11 @@ class sArticles
             $active = 1;
         }
 
+        $sBADWORDS = "#sex|porn|viagra|url\=|src\=|link\=#i";
+        if (preg_match($sBADWORDS, $sVoteComment)) {
+            return false;
+        }
+
         if (!empty($this->session['sArticleCommentInserts'][$article])) {
             $sql = '
                 DELETE FROM s_articles_vote WHERE id=?
@@ -526,8 +530,6 @@ class sArticles
             $id
         ));
 
-        $links = [];
-
         foreach ($getSupplier as $supplierKey => $supplierValue) {
             if (!Shopware()->Shop()->getDefault()) {
                 $getSupplier[$supplierKey] = $this->sGetTranslation($supplierValue, $supplierValue['id'], 'supplier');
@@ -537,29 +539,23 @@ class sArticles
                 $getSupplier[$supplierKey]["image"] = $mediaService->getUrl($supplierValue['image']);
             }
 
-            $supplierId = $supplierValue['id'];
             if ($id !== Shopware()->Shop()->getCategory()->getId()) {
-                $links[$supplierId] = [
+                $query = array(
                     'sViewport' => 'cat',
                     'sCategory' => $id,
                     'sPage' => 1,
-                    'sSupplier' => $supplierId
-                ];
+                    'sSupplier' => $supplierValue["id"]
+                );
             } else {
-                $links[$supplierId] = [
+                $query = array(
                     'controller' => 'listing',
                     'action' => 'manufacturer',
-                    'sSupplier' => $supplierId
-                ];
+                    'sSupplier' => $supplierValue["id"]
+                );
             }
-        }
 
-        $seoUrls = Shopware()->Container()->get('router')->generateList($links);
-        foreach ($getSupplier as &$supplier) {
-            $id = $supplier['id'];
-            if (array_key_exists($id, $seoUrls)) {
-                $supplier['link'] = $seoUrls[$id];
-            }
+            $getSupplier[$supplierKey]["link"] = Shopware()->Config()->get('baseFile')
+                .'?'.http_build_query($query, '', '&');
         }
 
         return $getSupplier;
@@ -707,8 +703,6 @@ class sArticles
         $criteria->limit($sLimitChart);
 
         $criteria->addSorting(new PopularitySorting(SortingInterface::SORT_DESC));
-
-        $criteria->setFetchCount(false);
 
         $result = $this->searchService->search($criteria, $context);
         $articles = $this->legacyStructConverter->convertListProductStructList($result->getProducts());
@@ -1231,8 +1225,8 @@ class sArticles
             return [];
         }
 
-        $hideNoInStock = $this->config->get('hideNoInStock');
-        if ($hideNoInStock && !$product->isAvailable()) {
+        $hideNoInstock = $this->config->get('hideNoInstock');
+        if ($hideNoInstock && !$product->isAvailable()) {
             return [];
         }
 
@@ -1477,8 +1471,6 @@ class sArticles
             default:
                 $criteria->addSorting(new ReleaseDateSorting(SortingInterface::SORT_DESC));
         }
-
-        $criteria->setFetchCount(false);
 
         $result = $this->productNumberSearch->search($criteria, $context);
 
@@ -2087,9 +2079,7 @@ class sArticles
             $mainKey = 0;
 
             if (empty($sCombination)) {
-                if (!empty($sArticle["image"]["res"]["description"])) {
-                    $sArticle["image"]["description"] = $sArticle["image"]["res"]["description"];
-                }
+                $sArticle["image"]["description"] = $sArticle["image"]["res"]["description"];
                 $sArticle["image"]["relations"] = $sArticle["image"]["res"]["relations"];
                 foreach ($sArticle["sConfigurator"] as $key => $group) {
                     foreach ($group["values"] as $key2 => $option) {
@@ -2299,7 +2289,7 @@ class sArticles
             'sArticles'       => $articles,
             'criteria'        => $criteria,
             'facets'          => $searchResult->getFacets(),
-            'sPage'           => (int) $request->getParam('sPage', 1),
+            'sPage'           => $request->getParam('sPage', 1),
             'pageSizes'       => $pageSizes,
             'sPerPage'        => $criteria->getLimit(),
             'sNumberArticles' => $searchResult->getTotalCount(),
@@ -2334,25 +2324,9 @@ class sArticles
 
             $convertedConfiguratorPrice = $this->legacyStructConverter->convertConfiguratorPrice($product, $configurator);
             $data = array_merge($data, $convertedConfiguratorPrice);
-
-            // generate additional text
-            if (!empty($selection)) {
-                $this->additionalTextService->buildAdditionalText($product, $this->contextService->getShopContext());
-                $data['additionaltext'] = $product->getAdditional();
-            }
-
-            if ($this->config->get('forceArticleMainImageInListing') && $configurator->getType() !== ConfiguratorService::CONFIGURATOR_TYPE_STANDARD && empty($selection)) {
-                $data['image'] = $this->legacyStructConverter->convertMediaStruct($product->getCover());
-                $data['images'] = [];
-                foreach ($product->getMedia() as $image) {
-                    if ($image->getId() !== $product->getCover()->getId()) {
-                        $data['images'][] = $this->legacyStructConverter->convertMediaStruct($image);
-                    }
-                }
-            }
         }
 
-        $data = array_merge($data, $this->getLinksOfProduct($product, $categoryId, !empty($selection)));
+        $data = array_merge($data, $this->getLinksOfProduct($product, $categoryId));
 
         $data["articleName"] = $this->sOptimizeText($data["articleName"]);
         $data["description_long"] = htmlspecialchars_decode($data["description_long"]);
@@ -2379,11 +2353,10 @@ class sArticles
      * Creates different links for the product like `add to basket`, `add to note`, `view detail page`, ...
      *
      * @param StoreFrontBundle\Struct\ListProduct $product
-     * @param int $categoryId
-     * @param bool $addNumber
+     * @param null $categoryId
      * @return array
      */
-    private function getLinksOfProduct(StoreFrontBundle\Struct\ListProduct $product, $categoryId, $addNumber)
+    private function getLinksOfProduct(StoreFrontBundle\Struct\ListProduct $product, $categoryId = null)
     {
         $baseFile = $this->config->get('baseFile');
         $context = $this->contextService->getShopContext();
@@ -2392,13 +2365,7 @@ class sArticles
         if ($categoryId) {
             $detail .= '&sCategory=' . $categoryId;
         }
-
         $rewrite = Shopware()->Modules()->Core()->sRewriteLink($detail, $product->getName());
-
-        if ($addNumber) {
-            $rewrite .= strpos($rewrite, '?') !== false ? '&' : '?';
-            $rewrite .= 'number=' . $product->getNumber();
-        }
 
         $basket = $baseFile . "?sViewport=basket&sAdd=" . $product->getNumber();
         $note = $baseFile . "?sViewport=note&sAdd=" . $product->getNumber();
